requirements.txt
Django==5.0
python-dotenv==1.0.0
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap4==2023.1
django-ckeditor==6.7.0
whitenoise==6.6.0
reportlab==4.0.8
python-slugify==8.0.1
beautifulsoup4==4.12.2
cffi==1.16.0
html5lib==1.1
cssselect2==0.7.0
tinycss2==1.2.1
pydyf==0.8.0 